Fix deprecated pygments warning
On the last Jekyll build of the site pages, the following warning was thrown:

You are attempting to use the 'pygments' highlighter, which is currently unsupported on GitHub Pages. Your site will use 'rouge' for highlighting instead. To suppress this warning, change the 'highlighter' value to 'rouge' in your '_config.yml' and ensure the 'pygments' key is unset. For more information, see https://help.github.com/en/articles/page-build-failed-config-file-error#fixing-highlighting-errors.

Updating the _config.yml to address the issue.